+++ changelog.md
@@ -7,6 +7,8 @@ v4.0.7
 * Changed research projects to use the Forge milk tag instead of just the milk bucket
 * Added new theorycrafting project to the pool
 * Made research projects refund container items, such as empty milk buckets
+* Increased base processing speed of most magitech devices
+* Increased total mana capacity of the Dissolution Chamber
 
 v4.0.6
 ------